The flow is, we first match antecedent using type match, if true, we instantiate the consequent using the matched graph concepts and then match consequent using text label match, if not (that means we need to apply this rule), we can add consequent to the graph. For example, if a graph has two sets of subgraphs with same structure, the rule will be applied twice instead of once. this is an improvement over previous versions.
While the method addAugmentedGraphObjects is copying the instantiated consequent to the original CG, we need to in the consequent, look for coref linked concepts and concepts already exist in the original CG. We modified method addAugmentedGraphObjects to make sure already exisitent concepts will not be copied again.

Modified removeFromGraph method in Graph class, when a concept is removed, it is also removed from conceptHashStore and equivalentConceptHashStore (use iterator remove rather than arraylist remove to avoid current modification exception)

Modified insertInCharGerGraph in Graph class line 624 is changed, when a concept is added, we also add it into equivalentConceptHashStore, we are not adding it to the sameTypeConceptHashStore since that may cause the rule to be applied over and over during one inference action.

generateEquivalentConceptSubgraphs